refactor: :recycle: move styles to own file
