+++ src/seedcase_flower/__init__.py
@@ -2,7 +2,16 @@
 
 from .cli import build, view
 from .config import Config
-from .internals import BuildStyle
 from .section import Content, Mode, Section
+from .styles import BuildStyle, ViewStyle
 
-__all__ = ["build", "view", "Config", "BuildStyle", "Content", "Mode", "Section"]
+__all__ = [
+    "build",
+    "view",
+    "Config",
+    "BuildStyle",
+    "ViewStyle",
+    "Content",
+    "Mode",
+    "Section",
+]
