+++ tests/test_cli.py
@@ -5,7 +5,7 @@ import json
 from pytest import fixture, mark
 
 from seedcase_flower.cli import build, view
-from seedcase_flower.internals import BuildStyle
+from seedcase_flower.styles import BuildStyle
 
 
 # Create a file at tmp_path that is automatically cleaned up after tests finish
